requirements.txt

python-telegram-bot==20.8
ccxt
python-dotenv
groq
